+++ TempleRunnerAndroid/app/src/main/java/com/sar/templerunner_android/GameLogic/SimpleCoin.java
@@ -2,7 +2,9 @@ package com.sar.templerunner_android.GameLogic;
 
 import android.content.res.Resources;
 
-public class SimpleCoin extends Coin{
+import com.sar.templerunner_android.GameLogic.coins.Coin;
+
+public class SimpleCoin extends Coin {
     public SimpleCoin(int x, int y, int coinSize, int screenY, Resources res) {
         super(x, y, coinSize, screenY, res);
     }
