+++ TempleRunnerAndroid/app/src/main/java/com/sar/templerunner_android/GameLogic/blocks/Block.java
@@ -1,4 +1,4 @@
-package com.sar.templerunner_android.GameLogic;
+package com.sar.templerunner_android.GameLogic.blocks;
 
 import android.content.res.Resources;
 import android.graphics.Bitmap;
@@ -9,6 +9,7 @@ import android.graphics.Point;
 import android.graphics.Rect;
 import android.os.DropBoxManager;
 
+import com.sar.templerunner_android.GameLogic.Player;
 import com.sar.templerunner_android.R;
 import com.sar.templerunner_android.Util.CurrentDifficulty;
