+++ TempleRunnerAndroid/app/src/main/java/com/sar/templerunner_android/GameLogic/blocks/SideBranch.java
@@ -1,13 +1,14 @@
-package com.sar.templerunner_android.GameLogic;
+package com.sar.templerunner_android.GameLogic.blocks;
 
 import android.content.res.Resources;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
 import android.graphics.Canvas;
 
+import com.sar.templerunner_android.GameLogic.Player;
 import com.sar.templerunner_android.R;
 
-public class RoadRock extends Block{
+public class SideBranch extends Block{
     private Bitmap obstacle;
     /**
      * screen on the x positions
@@ -22,9 +23,9 @@ public class RoadRock extends Block{
      * @param screenY   taille de l'ecan vers le bas
      * @param res
      */
-    public RoadRock(int x, int y, int blockSize, int screenY, Resources res) {
+    public SideBranch(int x, int y, int blockSize, int screenY, Resources res) {
         super(x, y, blockSize, screenY, res);
-        obstacle = BitmapFactory.decodeResource(res, R.drawable.rocher);
+        obstacle = BitmapFactory.decodeResource(res, R.drawable.sidebranch);
         obstacle=Bitmap.createScaledBitmap(obstacle, blockSize, blockSize, true);
     }
