+++ TempleRunnerAndroid/app/src/main/java/com/sar/templerunner_android/GameLogic/blocks/SimpleRoad.java
@@ -1,6 +1,8 @@
-package com.sar.templerunner_android.GameLogic;
+package com.sar.templerunner_android.GameLogic.blocks;
 
 import android.content.res.Resources;
+
+import com.sar.templerunner_android.GameLogic.Player;
 
 
 public class SimpleRoad extends Block {
