+++ TempleRunnerAndroid/app/src/main/java/com/sar/templerunner_android/GameLogic/blocks/branch.java
@@ -1,13 +1,14 @@
-package com.sar.templerunner_android.GameLogic;
+package com.sar.templerunner_android.GameLogic.blocks;
 
 import android.content.res.Resources;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
 import android.graphics.Canvas;
 
+import com.sar.templerunner_android.GameLogic.Player;
 import com.sar.templerunner_android.R;
 
-public class SideBranch extends Block{
+public class branch extends Block {
     private Bitmap obstacle;
     /**
      * screen on the x positions
@@ -22,15 +23,14 @@ public class SideBranch extends Block{
      * @param screenY   taille de l'ecan vers le bas
      * @param res
      */
-    public SideBranch(int x, int y, int blockSize, int screenY, Resources res) {
+    public branch(int x, int y, int blockSize, int screenY, Resources res) {
         super(x, y, blockSize, screenY, res);
-        obstacle = BitmapFactory.decodeResource(res, R.drawable.sidebranch);
+        obstacle = BitmapFactory.decodeResource(res, R.drawable.branch);
         obstacle=Bitmap.createScaledBitmap(obstacle, blockSize, blockSize, true);
     }
 
     @Override
     public boolean detectCollision(Player p) {
-        //TO DO
         return false;
     }
